attempt in fixing dupe glitch
It appears that all players have the value of ModReferences.ThisPlayerPacked of 0. every player should have an unique id.

Clients upon pressing e on an item send a request to the host to give them an item. If the item exists for the host, it sends a command to give player with some ID some item.
Since every player has the same id, all player get the item.

The ideal way would be to find a variable that every player has different. If im desperate ill use players name. Through i would like to avoid this solution.